+++ src/servo/es08a.c
@@ -86,7 +86,7 @@ upm_result_t es08a_set_angle(es08a_context dev, int32_t angle){
 
 upm_result_t es08a_calc_pulse_travelling(const es08a_context dev,
                                          int32_t* ret_val, int32_t value){
-    if (value > dev->max_pulse_width) {
+    if (value > (int)dev->max_pulse_width) {
         return dev->max_pulse_width;
     }
